Creating a LogMessage Interface (named LogMessage) and refactoring the previous LogMessage class to LogMessageImplementation; LogMessageImplementation
Using the Interface where applicable.

Creating a TimeStampValidatorMockedTest, which uses Mocked LogMessages, only having a timestamp and a signature counter for testing.
