+++ src/main/java/de/konfidas/ttc/messages/TransactionLogMessage.java
@@ -7,12 +7,13 @@ import org.bouncycastle.asn1.*;
 
 import java.io.IOException;
 import java.math.BigInteger;
+import java.util.Collection;
 import java.util.List;
 import java.util.ListIterator;
 
 
 
-public class TransactionLogMessage extends LogMessage {
+public class TransactionLogMessage extends LogMessageImplementation {
     String operationType;
     String clientID;
     byte[] processData;
@@ -86,7 +87,7 @@ public class TransactionLogMessage extends LogMessage {
 //        void parseCertifiedDataType(ByteArrayOutputStream dtbsStream, Enumeration<ASN1Primitive> asn1Primitives) throws IOException, LogMessage.CertifiedDataTypeParsingException, ExtendLengthValueExceedsInteger {
         super.parseCertifiedDataType(dtbsStream,logMessageAsASN1List,logMessageIterator);
         if(this.certifiedDataType != oid.id_SE_API_transaction_log){
-            throw new LogMessage.CertifiedDataTypeParsingException("Invalid Certified Data Type, expected id_SE_API_transaction_log but found "+this.certifiedDataType.getName(), null);
+            throw new LogMessageImplementation.CertifiedDataTypeParsingException("Invalid Certified Data Type, expected id_SE_API_transaction_log but found "+this.certifiedDataType.getName(), null);
         }
     }
 
@@ -275,4 +276,23 @@ return;
         super.checkContent();
     }
 
+    @Override
+    public int getVersion() {
+        return this.version;
     }
+
+    @Override
+    public oid getCertifiedDataType() {
+        return this.certifiedDataType;
+    }
+
+    @Override
+    public Collection<ASN1Primitive> getSignatureAlgorithmParameters() {
+        return this.signatureAlgorithmParameters;
+    }
+
+    @Override
+    public byte[] getSeAuditData() {
+        return this.seAuditData;
+    }
+}
