+++ src/main/java/de/konfidas/ttc/validation/TimeStampValidator.java
@@ -2,6 +2,7 @@ package de.konfidas.ttc.validation;
 
 import de.konfidas.ttc.exceptions.LogMessageValidationException;
 import de.konfidas.ttc.exceptions.ValidationException;
+import de.konfidas.ttc.messages.LogMessageImplementation;
 import de.konfidas.ttc.messages.LogMessage;
 import de.konfidas.ttc.messages.SystemLogMessage;
 import de.konfidas.ttc.messages.logtime.LogTime;
@@ -25,7 +26,7 @@ public class TimeStampValidator implements Validator{
         LinkedList<ValidationException> result = new LinkedList<>();
 
         ArrayList<LogMessage> messages = new ArrayList<>(tar.getLogMessages());
-        messages.sort(new LogMessage.SignatureCounterComparator());
+        messages.sort(new LogMessageImplementation.SignatureCounterComparator());
 
         String serial;
