+++ src/main/java/de/konfidas/ttc/validation/TransactionCounterValidator.java
@@ -2,7 +2,7 @@ package de.konfidas.ttc.validation;
 
 import de.konfidas.ttc.exceptions.LogMessageValidationException;
 import de.konfidas.ttc.exceptions.ValidationException;
-import de.konfidas.ttc.messages.LogMessage;
+import de.konfidas.ttc.messages.LogMessageImplementation;
 import de.konfidas.ttc.messages.TransactionLogMessage;
 import de.konfidas.ttc.tars.LogMessageArchive;
 
@@ -25,11 +25,11 @@ public class TransactionCounterValidator implements Validator {
     public Collection<ValidationException> validate(LogMessageArchive tar){
         ArrayList<ValidationException> result = new ArrayList<>();
 
-        ArrayList<LogMessage> msgs = new ArrayList(tar.getLogMessages());
-        msgs.sort(new LogMessage.SignatureCounterComparator());
+        ArrayList<LogMessageImplementation> msgs = new ArrayList(tar.getLogMessages());
+        msgs.sort(new LogMessageImplementation.SignatureCounterComparator());
 
 
-        for(LogMessage msg : msgs){
+        for(LogMessageImplementation msg : msgs){
             if(msg instanceof TransactionLogMessage){
                 result.addAll(updateState((TransactionLogMessage) msg));
             }
